AgriVisionAI-main/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
gunicorn
tensorflow-cpu==2.20.0
numpy==1.24.3
requests
Pillow
